+++ crates/digital_twin_glue/src/control/rocket.rs
@@ -1,5 +1,5 @@
 use crate::control::state::MissileState;
 
 pub trait Rocket {
-    fn update(&mut self, state: &MissileState, dt: f32) -> MissileState;
+    fn update(&mut self, state: &MissileState, dt: f64) -> MissileState;
 }
